requirements.txt
telebot
pydub
mutagen
azure.cognitiveservices.speech
openai
ffmpeg
websockets
prompt_toolkit
rich
emoji
httpx